src/requirements.txt
flask
asyncio-nats-client

